+++ infra/tf/modules/write_secret/vars.tf
@@ -1,0 +1,11 @@
+variable "triggers" {
+	type = map(string)
+}
+
+variable "path" {
+	type = string
+}
+
+variable "value" {
+	type = string
+}